Выполнены основные требования к работе, создан мидлвар для проверки авторизации, добавлен .env для хранения секретных ключей, также изменены контроллеры и роуты

+++ controllers/cards.js
@@ -7,15 +7,22 @@ const readCards = (req, res) => {
 };
 const createCard = (req, res) => {
   const { name, link } = req.body;
-  Card.create({ name, link, owner: req.user._id })
+  Card.create({ name, link, owner: req.user.id })
     .then((card) => res.status(200).send({ data: card }))
     .catch(() => res.status(500).send({ message: 'Произошла ошибка' }));
 };
 const deleteCard = (req, res) => {
-  Card.findByIdAndRemove(req.params.id)
+  Card.findById(req.params.id)
     .then((card) => {
       if (card) {
-        res.status(200).send({ data: card });
+        if (req.user.id == card.owner) {
+          Card.findByIdAndRemove(req.params.id)
+            .then(() => {
+              res.status(200).send({ data: card });
+            });
+        } else {
+          res.status(403).send({ message: 'Недостаточно прав для удаления карточки' });
+        }
       } else {
         res.status(404).send({ message: 'Такой карточки нет' });
       }
